Added -i and -p parameters to specify the ip and port on which to serve the web viewer. Separated the web server to a separate file in the utils directory.
