+++ requirements.txt
@@ -2,7 +2,7 @@ defusedxml==0.7.1
 ftfy==6.1.3
 nest_asyncio==1.5.8
 pyshark==0.6
-Requests==2.31.0
+Requests==2.32.0
 pytest==7.4.2
 Flask==3.0.3
 pandas==1.5.3
